Fix rescan form not showing on startup
